+++ apps/example/src/test.view.ts
@@ -1,67 +1,106 @@
-import { View, renderToString } from "@dlightjs/dlight"
+import { View, renderToString, Children, _ } from "@dlightjs/dlight"
 import { css, div } from "@dlightjs/easy-css"
-import { button, Env, htmlTag, Prop, required, SubView } from "@dlightjs/types"
+import { button, Env, htmlTag, Prop, required } from "@dlightjs/types"
 import { HStack, Route, RouterSpace, VStack } from "@dlightjs/components"
 import { MarkitView, addBlockRule } from "@dlightjs/markit"
+import { ForwardProp, Func } from "@dlightjs/decorators"
 
-class NNN extends View {
-  @Env path: any
+@ForwardProp
+@View
+class NNN {
   Body() {
-    div(this.path)
+    div()
+      .forwardProps()
   }
 }
 
-class TestView extends View {
+const Default = (() => {}) as any
+@View
+class JJ {
+  @Default @Prop qushifafe = required
+  @Children hh = required
+
+  willMount() {
+  }
+
+  Body() {
+    div(this.qushifafe)
+    _(this.hh)
+  }
+}
+
+function OK() {
+  console.log("jj")
+  this.loading = true
+  setTimeout(() => {
+    this.count++
+    this.loading = false
+  }, 1000)
+}
+
+function OK2() {
+  console.log("oo")
+  setInterval(() => {
+    this.count++
+  }, 1000)
+}
+@Func(OK)
+@Func(OK2)
+@View
+class TestView {
   count = 5
-
+  loading = false
   Body() {
-    HStack()
-    {
-      div("hh")
-      div("jj")
+    if (this.loading) {
+      div("shitttt")
     }
-    RouterSpace()
+    JJ("xefe")
     {
-      Route("hello")
-      {
-        div("hello")
-        NNN()
-      }
+      div("shit")
     }
+    div(this.count)
+    button("+")
+      .onclick(() => {
+        this.count++
+      })
+    NNN("hhh")
+      .style({
+        color: "red"
+      })
   }
 }
 
-export class TestMarkit extends View {
-  testMDString = `
-Search 🌟 in doc for important concepts and performance results.
+// export class TestMarkit extends View {
+//   testMDString = `
+// Search 🌟 in doc for important concepts and performance results.
 
-# Quick start
+// # Quick start
 
-DLight uses [vite](https://vitejs.dev/) to construct its apps. We mainly use [this vite plugin](https://www.npmjs.com/package/vite-plugin-dlight-transpiler) to transpile jsx/jsd file into pure js code.
+// DLight uses [vite](https://vitejs.dev/) to construct its apps. We mainly use [this vite plugin](https://www.npmjs.com/package/vite-plugin-dlight-transpiler) to transpile jsx/jsd file into pure js code.
 
-Three ways to try DLight.js out.
+// Three ways to try DLight.js out.
 
-* Use CLI to build a dlight app. (**This feature is still in development.**)
+// * Use CLI to build a dlight app. (**This feature is still in development.**)
 
-\`\`\`shell
-npm install -g @dlightjs/cli
-create-dlight-app my-first-dlight-app
-\`\`\`
+// \`\`\`shell
+// npm install -g @dlightjs/cli
+// create-dlight-app my-first-dlight-app
+// \`\`\`
 
-* Clone this repo https://github.com/dlight-js/dlight-vite-template for a quick start.
-* 🌟 Play around in [codesandbox](https://codesandbox.io/p/sandbox/dlight-vite-quickstart-4tgogd)
-  `
+// * Clone this repo https://github.com/dlight-js/dlight-vite-template for a quick start.
+// * 🌟 Play around in [codesandbox](https://codesandbox.io/p/sandbox/dlight-vite-quickstart-4tgogd)
+//   `
 
-  getAst = (ast: any) => {
-    console.log(ast)
-  }
+//   getAst = (ast: any) => {
+//     console.log(ast)
+//   }
 
-  Body() {
-    MarkitView(this.testMDString)
-      .getAst(this.getAst)
-  }
-}
+//   Body() {
+//     MarkitView(this.testMDString)
+//       .getAst(this.getAst)
+//   }
+// }
 
-console.log(renderToString(TestView))
+// console.log(renderToString(TestView))
 
 export default TestView