+++ packages/core/types/src/test.ts
@@ -1,5 +1,5 @@
 import { View } from "@dlightjs/dlight"
-import { type div, type Pretty, type Typed, type UnTyped } from "./index"
+import { Default, type div, type Pretty, type Typed, type UnTyped } from "./index"
 
 const VV = View as any as (typeof View & ((...args: any) => any))
 /**
@@ -7,7 +7,7 @@ const VV = View as any as (typeof View & ((...args: any) => any))
 */
 interface Props {
   /** jfjsf */
-  a: string
+  a?: Default<string>
   b: string
   c?: number
   _$content?: string
@@ -23,10 +23,13 @@ class JJClass extends VV implements Props {
   }
 }
 const t = JJClass as Pretty as Typed<Props>
+// type j = DefaultKeyName<Props>
 t()
   .c(1)
+  .
 const g: Typed<UnTyped<typeof t>, HTMLDivElement> = null as any
 g()
   .a("1")
   .b("2")
   .c(3)
+  .