+++ apps/test/CHANGELOG.md
@@ -1,5 +1,25 @@
 # dlight-test
 
+## 0.0.8
+
+### Patch Changes
+
+- Updated dependencies
+  - vite-plugin-dlight@0.9.28
+  - @dlightjs/components@0.9.33
+  - @dlightjs/transpiler@0.9.28
+
+## 0.0.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @dlightjs/dlight@0.9.29
+  - @dlightjs/types@0.9.29
+  - @dlightjs/components@0.9.32
+  - @dlightjs/transpiler@0.9.27
+  - vite-plugin-dlight@0.9.27
+
 ## 0.0.6
 
 ### Patch Changes