+++ packages/core/babel-preset-dlight/CHANGELOG.md
@@ -1,5 +1,17 @@
 # babel-preset-dlight
 
+## 0.9.28
+
+### Patch Changes
+
+- feat: change to decorators family
+
+## 0.9.27
+
+### Patch Changes
+
+- feat: add advanced dx
+
 ## 0.9.26
 
 ### Patch Changes