+++ packages/core/types/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @dlightjs/types
 
+## 0.9.29
+
+### Patch Changes
+
+- feat: add advanced dx
+- Updated dependencies
+  - @dlightjs/dlight@0.9.29
+
 ## 0.9.28
 
 ### Patch Changes