+++ packages/ecosystem/vite-plugin-dlight/CHANGELOG.md
@@ -1,5 +1,20 @@
 # vite-plugin-dlight
 
+## 0.9.28
+
+### Patch Changes
+
+- feat: change to decorators family
+- Updated dependencies
+  - babel-preset-dlight@0.9.28
+
+## 0.9.27
+
+### Patch Changes
+
+- Updated dependencies
+  - babel-preset-dlight@0.9.27
+
 ## 0.9.26
 
 ### Patch Changes